FIX: Reorganize meshdata
