+++ vispy/geometry/__init__.py
@@ -8,10 +8,12 @@ This module implements classes and methods for handling geometric data.
 
 from __future__ import division
 
-__all__ = ['Triangulation', 'PolygonData', 'Rect', 'create_cube']
+__all__ = ['MeshData', 'PolygonData', 'Rect', 'Triangulation', 'create_cube',
+           'create_cylinder', 'create_sphere']
 
 from .polygon import PolygonData  # noqa
+from .meshdata import MeshData  # noqa
 from .rect import Rect  # noqa
 from .triangulation import Triangulation  # noqa
 from .calculations import _calculate_normals, _fast_cross_3d  # noqa
-from .generation import create_cube  # noqa
+from .generation import create_cube, create_cylinder, create_sphere  # noqa
